+++ CHANGELOG.md
@@ -9,5 +9,6 @@
 * 验证码单独成为一个模块hutool-captcha
 * 新增NamedThreadFactory
 * 新增BufferUtil
+* POI新增StyleUtil，StyleSet新增方法可设置背景、边框等样式
 
 ### Bug修复